+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 ### Fixed
 - proper calculating data for production that takes more days
 - proper displaying group and workplace data on initial load
+- added proper locales
 
 
 ## [2021.1.1.20] - 2021-01-20