+++ CHANGELOG.md
@@ -14,6 +14,13 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.1.1.25] - 2021-01-25
+
+### Added
+- standard charts workplace select, with proper functionality
+- standard charts date from and date to select, with proper functionality
+- timeline chart version 1
+
 ## [2021.1.1.21] - 2021-01-21
 
 ### Fixed